move api calls into "browser" folder
