+++ packages/app/src/api/browser/types/pulse.d.ts
