+++ packages/app/src/components/WeeklyReports/ReportsPage.tsx
@@ -2,13 +2,13 @@ import dayjs from 'dayjs'
 import { Box, Divider, Stack, Typography } from '@mui/material'
 import { DateHeader } from '../Home/DateHeader'
 import { useState } from 'react'
-import { useGetWeeklyReport } from '../../api/report.api'
 import { ProjectScore } from './ProjectScore'
 import { GrowthScore } from './GrowthScore'
 import { DeepWorkScore } from './DeepWorkScore'
 import { ActiveHoursScore } from './ActiveHoursScore'
 import { Logo } from '../common/Logo/Logo'
 import InfoOutlinedIcon from '@mui/icons-material/InfoOutlined'
+import { useGetWeeklyReport } from '../../api/browser/report.api'
 
 export const ReportsPage = () => {
   const period = 'week'
