+++ packages/app/src/components/WeeklyReports/ScoreHeader.tsx
@@ -1,5 +1,5 @@
 import { Box, Typography } from '@mui/material'
-import { getColorForRating } from '../../api/services/report.service'
+import { getColorForRating } from '../../api/browser/services/report.service'
 
 interface Props {
   title: string
