+++ packages/app/src/extensions/SqlSandbox/sandbox.api.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query'
-import { sqlQueryFn } from '../../api/services/query.service'
+import { sqlQueryFn } from '../../api/browser/services/query.service'
 
 export const useRunSql = () => {
   return useMutation({
